refactor: 패키지 구조화 진행
