+++ src/main/java/lotto/utils/validator/WinningNumbersValidator.java
@@ -1,8 +1,8 @@
 package lotto.utils.validator;
 
-import static lotto.utils.validator.ErrorMessageConstants.COMMA_POSITION_ERROR_MESSAGE;
-import static lotto.utils.validator.ErrorMessageConstants.ERROR_MESSAGE_HEAD;
-import static lotto.utils.validator.ErrorMessageConstants.NUMBER_FORMAT_ERROR_MESSAGE;
+import static lotto.ErrorMessageConstants.COMMA_POSITION_ERROR_MESSAGE;
+import static lotto.ErrorMessageConstants.ERROR_MESSAGE_HEAD;
+import static lotto.ErrorMessageConstants.NUMBER_FORMAT_ERROR_MESSAGE;
 
 import java.util.ArrayList;
 import java.util.List;
