+++ src/main/java/lotto/view/OutView.java
@@ -1,10 +1,12 @@
-package lotto;
+package lotto.view;
 
 import java.sql.SQLOutput;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import lotto.domain.Lotto;
+import lotto.domain.Ranking;
 
 public class OutView {
